Fix file_length linter warning
