+++ Sources/TokamakCore/Shapes/Path/PathSizing.swift
